Added the name of the server to the subject line of error report emails that are sent to the software tester email address.
Fixed bug that was causing a sql error b/c of a lack of sample id's in the where clause.  Now we check to make sure that there are sample ids in the list before running the query.  This was occurring in the delExpFileAndNotify function.

Put a check in to make sure that the toAddress of the delExpAndNotify function email is valid.  If it is not we print an error to the console and don't try to send the email.